function_load/requirements.txt
azure-functions==1.11.0
azure-storage-blob==12.24.0
